pyproject.toml
[project]
name = "libre-linkup-py"
version = "0.2.4"
description = "Unofficial libre linkup python client"
authors = [{ name = "Sam Purkis", email = "sam.purkis@hotmail.co.uk" }]
dependencies = ["python-dotenv>=1.0.1", "requests>=2.31.0", "pytz>=2024.1"]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
dev-dependencies = [
    "mypy>=1.8.0",
    "pytest>=8.0.2",
    "pydantic>=2.6.3",
    "pytest-dependency>=0.6.0",
    "plotly>=5.19.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/libre_link_up"]
